refactor(internal): move assertCapData into its own module
